requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Faker==32.1.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
kafka-python==2.0.2
pymongo==4.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
